+++ components/combobox/combobox.vue
@@ -30,9 +30,7 @@
       <slot
         name="list"
         :list-props="listProps"
-        :get-item-props="getItemProps"
-        :active-item-index="highlightIndex"
-        :set-highlight-index="setHighlightIndex"
+        :opened="onOpen"
       />
     </div>
   </div>
@@ -48,7 +46,8 @@ export default {
   mixins: [
     KeyboardNavigation({
       indexKey: 'highlightIndex',
-      listElementKey: 'listRef',
+      idKey: 'highlightId',
+      listElementKey: 'getListElement',
       afterHighlightMethod: 'afterHighlight',
       beginningOfListMethod: 'beginningOfListMethod',
       endOfListMethod: 'endOfListMethod',
@@ -101,6 +100,15 @@ export default {
 
   emits: ['select', 'escape', 'highlight'],
 
+  data () {
+    return {
+      // If the list is rendered at the root, rather than as a child
+      // of this component, this is the ref to that dom element. Set
+      // by the onOpen method.
+      outsideRenderedListRef: null,
+    };
+  },
+
   computed: {
     inputProps () {
       return {
@@ -120,21 +128,6 @@ export default {
       };
     },
 
-    listRef () {
-      return this.$refs.listWrapper;
-    },
-
-    /*
-     * These props are wrapped in a function that expects that an index is passed.
-     */
-    getItemProps () {
-      return (i) => ({
-        role: 'option',
-        // The ids have to be generated here since we use them for activeItemId.
-        id: this.getItemId(i),
-      });
-    },
-
     beginningOfListMethod () {
       return this.onBeginningOfList || this.jumpToEnd;
     },
@@ -147,28 +140,30 @@ export default {
       if (!this.showList || this.highlightIndex < 0) {
         return;
       }
-      return this.getItemId(this.highlightIndex);
+      return this.highlightId;
     },
 
     activeItemEl () {
-      return document.getElementById(this.activeItemId);
+      return this.getListElement().querySelector('#' + this.highlightId);
     },
   },
 
   watch: {
     showList () {
       // When the list's visibility changes reset the highlight index.
-      this.setHighlightIndex(0);
+      this.$nextTick(function () {
+        this.setInitialHighlightIndex();
+      });
     },
   },
 
   methods: {
-    clearHighlightIndex () {
-      this.setHighlightIndex(-1);
+    getListElement () {
+      return this.outsideRenderedListRef ?? this.$refs.listWrapper;
     },
 
-    getItemId (i) {
-      return `${this.listId}-item${i}`;
+    clearHighlightIndex () {
+      this.setHighlightIndex(-1);
     },
 
     afterHighlight () {
@@ -183,6 +178,17 @@ export default {
 
     onEscapeKey () {
       this.$emit('escape');
+    },
+
+    onOpen (open, contentRef) {
+      this.outsideRenderedListRef = contentRef;
+    },
+
+    setInitialHighlightIndex () {
+      if (this.showList) {
+        // When the list's is shown, reset the highlight index.
+        this.setHighlightIndex(0);
+      }
     },
   },
 };